Fix and refactor IT
