+++ integrationtesting/src/config.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Config:
+    client_url: str
+    server_url: str
+
+
+config: Config = Config(
+    client_url="http://localhost:3000",
+    server_url="http://localhost:8080"
+)
