+++ integrationtesting/src/test_sample.py
@@ -1,26 +1,14 @@
 import time
-from dataclasses import dataclass
 
 from selenium import webdriver
 from selenium.webdriver.chrome.webdriver import WebDriver
 from selenium.webdriver.common.by import By
 
-
-@dataclass
-class Config:
-    client_url: str
-    server_url: str
-
-
-config: Config = Config(
-    client_url="http://localhost:3000",
-    server_url="http://localhost:8080"
-)
+from config import config
 
 
 def test_register_parcel():
     options = webdriver.ChromeOptions()
-    options.headless = True
     driver: WebDriver = webdriver.Chrome(options=options)
     driver.implicitly_wait(5)
     driver.get(config.client_url)
@@ -28,9 +16,9 @@ def test_register_parcel():
     element.click()
 
     time.sleep(0.5)
-    element = driver.find_element(By.XPATH, "//button[contains(text(), 'Next')]")
+    element = driver.find_element(By.ID, "next-button-1")
     element.click()
 
     time.sleep(0.5)
-    element = driver.find_element(By.XPATH, "//button[contains(text(), 'Next')]")
+    element = driver.find_element(By.ID, "next-button-2")
     element.click()
