+++ integrationtesting/src/test_server.py
@@ -1,18 +1,5 @@
-from dataclasses import dataclass
-
 from server import ServerClient
-
-
-@dataclass
-class Config:
-    client_url: str
-    server_url: str
-
-
-config: Config = Config(
-    client_url="http://localhost:3000",
-    server_url="http://localhost:8080"
-)
+from config import config
 
 
 def test_scenario():
